@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 
-import {Observable} from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import {Observable,  of } from 'rxjs';
 
 import {Person} from '../models/person.model';
 import {PEOPLE} from '../mock/people';
