@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 
-import { Information } from '@app/core';
+import { Attribute } from '@app/core';
 
 
 
@@ -10,7 +10,7 @@ import { Information } from '@app/core';
   styleUrls: ['./information.component.scss']
 })
 export class InformationComponent implements OnInit {
-  @Input() info: Information;
+  @Input() info: Attribute;
   editing: boolean = false;
 
   constructor() { }
